% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sctrip.r
\name{sc_object.trip}
\alias{sc_object.trip}
\title{Convert trip to common form.}
\usage{
\method{sc_object}{trip}(x, ...)
}
\arguments{
\item{x}{trip object}
}
\description{
Convert trip to common form.
}
\examples{
library(trip)
sc_object(walrus818)
sc_coord(walrus818)
sc_path(walrus818)

PATH(walrus818)
}
