% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/path.r
\name{sc_path}
\alias{sc_path}
\title{Title}
\usage{
sc_path(data, ...)
}
\arguments{
\item{data}{input data}

\item{...}{columns that are nominated as vertex coordinates}
}
\value{
`sc::PATH`
}
\description{
Nominated columns will be used to find unique coordinates. All other columns belong on
the path-link-vertex data. Currently this is to find a single path from a table of data.
In future there will be methods for grouped data frames, and for other ways of declaring
groupings at the path and object level.
}
\examples{
data("Seatbelts", package= "datasets")
apath <- sc_path(as.data.frame(Seatbelts), front, rear, kms)
sc::PRIMITIVE(apath)
sc_path(aurora)
}
